more progress on local database integration, documentation/comment changes
I'm trying to get to the point where I could bring another Java developer onto the project and have it be easy to start developing for. I don't really have any desire to make this a collaborative project, but I think that being able to bring on another dev is a good place to be as far as documentation and commented code goes. Also will probably need to clean up the local and AWS DB classes; there's too many static variables and helper functions imo